flatbond/requirements.txt
Django==1.11.6
djangorestframework==3.9.1
pytz==2018.9
